pyproject.toml
[tool.poetry]
name = "ue-calendar-scraper"
version = "0.1.0"
description = ""
authors = ["jonas <jonasjreynolds@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
beautifulsoup4 = "^4.10.0"
google-api-python-client = "^2.28.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
